Update README.md
## Description
This project aims to provide a tool for predicting mental disorders based on demographic and symptomatic data. Mental health is a critical aspect of overall well-being, and early detection and intervention can significantly improve outcomes for individuals. By leveraging machine learning algorithms, this classifier offers a means to assist healthcare professionals in identifying potential mental health concerns in patients.

The provided dataset, while not disclosed due to privacy reasons, contains anonymized information collected from individuals diagnosed with various mental disorders. The attributes in the dataset encompass a wide range of factors, including age, gender, family history, and symptoms reported by the patients. 

This project serves as a starting point for developing a robust and accurate classifier that can be integrated into healthcare systems or used for research purposes. It encourages further exploration and refinement of machine learning techniques in the field of mental health diagnosis and treatment.